added enum Shape, updated abstract class Figure with shape as name

+++ src/main/java/core/basesyntax/figure/Squared.java
@@ -2,6 +2,6 @@ package core.basesyntax.figure;
 
 public interface Squared {
 
-    int getSquare();
+    double getSquare();
 
 }
